refactor(web): make Submissions subsection own the submission definition
Extract the submission-mode + milestone-tags controls into a dedicated
Submissions subsection whose definition control is always visible,
independent of the grading choice. Only the shim-dependent affordances
(milestone tags, the retrofit edit warning) stay gated on a built-in
autograder. Autograding and the submissions page consume this definition
rather than owning it.
