+++ web/src/locales/en.json
@@ -1193,10 +1193,14 @@
           "help": "Total points available for this manual assignment. Used to show each student's score out of the maximum."
         }
       },
+      "submissions": {
+        "heading": "Submissions",
+        "help": "What counts as a submission for this assignment — how the submissions page detects and counts each student's work. The autograder consumes this same definition."
+      },
       "submissionMode": {
-        "label": "Autograding trigger",
-        "help": "When the autograder runs. \"Every push\" grades each push to the default branch. \"On submit only\" grades only when a student submits (with gh student submit) or pushes a submit/* tag — regular pushes cost no Actions minutes, which matters for large classes.",
-        "editWarning": "Changing the trigger only affects repositories created from now on. Repositories students already accepted keep the old trigger until you update them — use \"Update autograding triggers\" on the submissions page or gh teacher assignment submission-mode. Students must pull after the update.",
+        "label": "What counts as a submission",
+        "help": "\"Every push\" counts each push to the default branch (excluding the initial commit) as a submission, and — with a built-in autograder — grades each one. \"On submit only\" counts a submission only when a student submits (with gh student submit) or pushes a submit/* tag; regular pushes cost no Actions minutes, which matters for large classes. The autograder uses this same definition to decide when it runs.",
+        "editWarning": "Changing this only affects repositories created from now on. Repositories students already accepted keep the old trigger until you update them — use \"Update autograding triggers\" on the submissions page or gh teacher assignment submission-mode. Students must pull after the update.",
         "choices": {
           "everyPush": "Every push (default)",
           "tag": "On submit only"
