+++ web/src/pages/assignments/sections/SubmissionGradingSection.tsx
@@ -1,25 +1,23 @@
 import { useTranslation } from "react-i18next"
-import { Alert, FormField, Input, Select, Textarea } from "@/components/ui"
-import {
-  parseSubmissionTags,
-  validateSubmissionTags,
-} from "@/util/submissionTags"
+import { Alert, FormField, Input, Select } from "@/components/ui"
 import { GRADING_MAX_POINTS_MIN } from "@/types/classroom"
 import type { AssignmentForm } from "../assignmentFormModel"
 import { deriveFormShape } from "../formShape"
 import type { SectionStatus } from "./sectionStatus"
 import { SectionCard } from "./SectionCard"
+import { SubmissionsSubsection } from "./SubmissionsSubsection"
 
 // Submission and Grading: how the assignment is graded (off / auto / manual,
-// with a manual max-points) and — when a built-in shim exists — what triggers
-// the autograder (every push vs. on submit only, plus milestone tags).
+// with a manual max-points) and — under the Submissions subsection — what
+// counts as a submission (the definition + milestone tags).
 //
-// The grading controls apply to ANY assignment (a bare or teacher-CI repo can
-// still be graded by hand), so the section always renders. The submission
-// trigger + milestone tags only matter with a built-in shim (they are mutually
-// exclusive with empty_repo / no_autograder on the wire), so those sub-controls
-// stay gated on showBuiltInConfig. Hidden fields keep their (default) values, so
-// the wire stays correct.
+// The grading controls and the submission DEFINITION both apply to ANY
+// assignment (a bare or teacher-CI repo can still be graded by hand and can
+// still declare what counts as a submission), so they always render. Only the
+// shim-dependent affordances inside the Submissions subsection (milestone tags,
+// the retrofit edit warning) stay gated on showBuiltInConfig, because a bare
+// repo or teacher-supplied CI has no shim to trigger. Hidden fields keep their
+// (default) values, so the wire stays correct.
 export function SubmissionGradingSection({
   form,
   edit,
@@ -39,21 +37,28 @@ export function SubmissionGradingSection({
     >
       <div className="flex flex-col gap-4">
         <GradingChoiceField form={form} edit={edit} />
-        {/* Submission trigger + milestone tags need a built-in shim; hidden for
-            a bare repo or teacher-supplied CI. */}
-        <form.Subscribe
-          selector={(state) => deriveFormShape(state.values).showBuiltInConfig}
-        >
-          {(showBuiltInConfig) =>
-            showBuiltInConfig ? (
-              <>
-                <div className="divider my-0" />
-                <SubmissionModeField form={form} edit={edit} />
-                <SubmissionTagsField form={form} edit={edit} />
-              </>
-            ) : null
-          }
-        </form.Subscribe>
+        <div className="divider my-0" />
+        <div className="flex flex-col gap-2">
+          <h3 className="text-sm font-semibold">
+            {t("assignments.form.submissions.heading")}
+          </h3>
+          <p className="text-sm text-base-content/70">
+            {t("assignments.form.submissions.help")}
+          </p>
+          <form.Subscribe
+            selector={(state) =>
+              deriveFormShape(state.values).showBuiltInConfig
+            }
+          >
+            {(showBuiltInConfig) => (
+              <SubmissionsSubsection
+                form={form}
+                edit={edit}
+                showBuiltInConfig={showBuiltInConfig}
+              />
+            )}
+          </form.Subscribe>
+        </div>
       </div>
     </SectionCard>
   )
@@ -177,152 +182,6 @@ function ManualMaxPointsField({ form }: { form: AssignmentForm }) {
                 {error}
               </p>
             ) : null}
-          </div>
-        )
-      }}
-    </form.Field>
-  )
-}
-
-function SubmissionModeField({
-  form,
-  edit,
-}: {
-  form: AssignmentForm
-  edit: boolean
-}) {
-  const { t } = useTranslation()
-  return (
-    <form.Field name="submission_mode">
-      {(field) => (
-        <div>
-          <FormField
-            htmlFor={field.name}
-            label={t("assignments.form.submissionMode.label")}
-            help={t("assignments.form.submissionMode.help")}
-          >
-            {({ id, describedById }) => (
-              <Select
-                id={id}
-                name={field.name}
-                className="w-full sm:max-w-xs"
-                aria-describedby={describedById}
-                value={field.state.value}
-                onBlur={field.handleBlur}
-                onChange={(e) =>
-                  field.handleChange(e.target.value as typeof field.state.value)
-                }
-              >
-                <option value="every-push">
-                  {t("assignments.form.submissionMode.choices.everyPush")}
-                </option>
-                <option value="tag">
-                  {t("assignments.form.submissionMode.choices.tag")}
-                </option>
-              </Select>
-            )}
-          </FormField>
-          {edit ? (
-            <form.Subscribe selector={(state) => state.values.submission_mode}>
-              {(mode) =>
-                mode !==
-                (form.options.defaultValues?.submission_mode ??
-                  "every-push") ? (
-                  <Alert tone="warning" role="status" className="mt-2 text-sm">
-                    <span>
-                      {t("assignments.form.submissionMode.editWarning")}
-                    </span>
-                  </Alert>
-                ) : null
-              }
-            </form.Subscribe>
-          ) : null}
-        </div>
-      )}
-    </form.Field>
-  )
-}
-
-function SubmissionTagsField({
-  form,
-  edit,
-}: {
-  form: AssignmentForm
-  edit: boolean
-}) {
-  const { t } = useTranslation()
-  return (
-    <form.Field name="submission_tags">
-      {(field) => {
-        const error = field.state.meta.errors[0] as string | undefined
-        return (
-          <div>
-            <FormField
-              htmlFor={field.name}
-              label={t("assignments.form.submissionTags.label")}
-              help={t("assignments.form.submissionTags.help")}
-            >
-              {({ id, describedById }) => (
-                <Textarea
-                  id={id}
-                  name={field.name}
-                  className="font-mono w-full sm:max-w-xs"
-                  rows={3}
-                  spellCheck={false}
-                  placeholder={"phase1\nphase2\ncomplete"}
-                  aria-describedby={describedById}
-                  value={field.state.value}
-                  onBlur={field.handleBlur}
-                  onChange={(e) => field.handleChange(e.target.value)}
-                />
-              )}
-            </FormField>
-            {error ? (
-              <p role="alert" className="mt-1.5 text-sm text-error">
-                {error}
-              </p>
-            ) : null}
-            <form.Subscribe selector={(state) => state.values.submission_tags}>
-              {(tags) => {
-                // Surface the same problems the save-time validator catches,
-                // live and ahead of save (the form only validates onSubmit).
-                // Priority, most to least actionable:
-                //   1. comma — the obvious wrong guess for a one-per-line
-                //      field; its own friendly hint since the generic charset
-                //      error wouldn't explain the real fix.
-                //   2. a hard validation error (bad charset, duplicate,
-                //      stacked quantifier) — the exact save-path message.
-                //   3. broad-glob caution / edit-retrofit warning — advisory.
-                const parsed = parseSubmissionTags(tags)
-                const hasComma = (tags ?? "").includes(",")
-                const validationError = validateSubmissionTags(parsed)
-                const broad = parsed.some(
-                  (p) => p.includes("*") || p.includes("+"),
-                )
-                const changed =
-                  edit &&
-                  tags !== (form.options.defaultValues?.submission_tags ?? "")
-                if (!hasComma && !validationError && !broad && !changed)
-                  return null
-                const message = hasComma
-                  ? t("assignments.form.submissionTags.commaHint")
-                  : (validationError ??
-                    (broad
-                      ? t("assignments.form.submissionTags.wildcardCaution")
-                      : t("assignments.form.submissionMode.editWarning")))
-                // Errors read as errors; the advisory cautions stay warning.
-                const isError = hasComma || Boolean(validationError)
-                return (
-                  <Alert
-                    tone={isError ? "error" : "warning"}
-                    role="status"
-                    className="mt-2 text-sm"
-                  >
-                    <span>{message}</span>
-                  </Alert>
-                )
-              }}
-            </form.Subscribe>
           </div>
         )
       }}
